pyproject.toml
[tool.poetry]
name = "squirrel-bot"
version = "1.0.0"
description = "A funny bot that only reacts with the squirrel emoji to each message"
authors = ["Leonardo Guedez <leonardoguedezsuazo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^1.7.3"
python-dotenv = "^0.20.0"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
